refactor: move tests in When_adding_outbox_messages into correctly named types
