+++ src/NServiceBus.RavenDB.Tests/Outbox/When_storing_outbox_messages.cs
@@ -13,7 +13,7 @@ namespace NServiceBus.RavenDB.Tests.Outbox
     using Raven.Client.Exceptions.Documents.Session;
 
     [TestFixture]
-    public class When_adding_outbox_messages : RavenDBPersistenceTestBase
+    public class When_storing_outbox_messages : RavenDBPersistenceTestBase
     {
         string testEndpointName = "TestEndpoint";
 
@@ -127,103 +127,6 @@ namespace NServiceBus.RavenDB.Tests.Outbox
                 var metadata = session.Advanced.GetMetadataFor(outboxRecord);
 
                 Assert.AreEqual(OutboxRecord.SchemaVersion, metadata[SchemaVersionExtensions.OutboxRecordSchemaVersionMetadataKey]);
-            }
-        }
-
-        [Test]
-        public async Task Should_update_dispatched_flag()
-        {
-            // arrange
-            var persister = new OutboxPersister(testEndpointName, CreateTestSessionOpener());
-            var context = new ContextBag();
-            var incomingMessageId = SimulateIncomingMessage(context).MessageId;
-            var outboxMessage = new OutboxMessage(incomingMessageId, new[] { new TransportOperation(incomingMessageId, default, default, default) });
-
-            using (var transaction = await persister.BeginTransaction(context))
-            {
-                await persister.Store(outboxMessage, transaction, context);
-                await transaction.Commit();
-            }
-
-            // act
-            await persister.SetAsDispatched(incomingMessageId, context);
-            WaitForIndexing();
-
-            // assert
-            using (var session = store.OpenAsyncSession())
-            {
-                var outboxRecord = await session.Query<OutboxRecord>().SingleOrDefaultAsync(record => record.MessageId == incomingMessageId);
-
-                Assert.NotNull(outboxRecord);
-                Assert.True(outboxRecord.Dispatched);
-            }
-        }
-
-        [Test]
-        public async Task Should_get_messages()
-        {
-            // arrange
-            var persister = new OutboxPersister(testEndpointName, CreateTestSessionOpener());
-            var context = new ContextBag();
-            var incomingMessageId = SimulateIncomingMessage(context).MessageId;
-
-            //manually store an OutboxRecord to control the OutboxRecordId format
-            using (var session = OpenAsyncSession())
-            {
-                var outboxRecord = new OutboxRecord
-                {
-                    MessageId = incomingMessageId,
-                    Dispatched = false,
-                    TransportOperations = new[] { new OutboxRecord.OutboxOperation { MessageId = incomingMessageId, } }
-                };
-
-                var outboxRecordId = "Outbox/TestEndpoint/" + incomingMessageId;
-
-                await session.StoreAsync(outboxRecord, outboxRecordId);
-                await session.SaveChangesAsync();
-            }
-
-            // act
-            var outboxMessage = await persister.Get(incomingMessageId, context);
-
-            // assert
-            Assert.NotNull(outboxMessage);
-            Assert.AreEqual(incomingMessageId, outboxMessage.MessageId);
-        }
-
-        [Test]
-        public async Task Should_set_messages_as_dispatched()
-        {
-            // arrange
-            var persister = new OutboxPersister(testEndpointName, CreateTestSessionOpener());
-            var context = new ContextBag();
-            var incomingMessageId = SimulateIncomingMessage(context).MessageId;
-
-            //manually store an OutboxRecord to control the OutboxRecordId format
-            using (var session = OpenAsyncSession())
-            {
-                await session.StoreAsync(
-                    new OutboxRecord
-                    {
-                        MessageId = incomingMessageId,
-                        Dispatched = false,
-                        TransportOperations = new[] { new OutboxRecord.OutboxOperation { MessageId = incomingMessageId, } }
-                    },
-                    "Outbox/TestEndpoint/" + incomingMessageId);
-
-                await session.SaveChangesAsync();
-            }
-
-            // act
-            await persister.SetAsDispatched(incomingMessageId, context);
-
-            // assert
-            using (var session = OpenAsyncSession())
-            {
-                var outboxRecord = await session.LoadAsync<OutboxRecord>("Outbox/TestEndpoint/" + incomingMessageId);
-
-                Assert.NotNull(outboxRecord);
-                Assert.True(outboxRecord.Dispatched);
             }
         }
